A lot going on behind the scenes! The interface is finally looking as it should.
The search bar will eventually double as a filter for any table that currently
being displated as well as a form that gets submitted and leads to the main search
page of the application.  The Farmer page can now accept parameters to query the
search engine directly reducing the load on the serverside.

Cheers!