I am changing the version number of the application to version 0.5.
This represents a point where the application can successfully be used but still requires some
refinement.

I am also pushing a commit to the travis file that will allow for automated pushes to heroku.
This should allow for development and production pushes to the environments.  Travis will now
push the successful builds to the relevant enviroments using an encrypted api key.

Matthew Budram