Big Changes!  I introduced a new table/collection in mongo that stores each report generated by the system.
This allows us to quickly refer to a previously generated report instead of attempting to regenerate a new one.

Also the report itself is NOT complete.  There are fields that are missing from the report itself.

The function to render a report is also imcomplete so look out for TODO markers in the code.  Code should be
commented properly and this was lacking for several commits.